refactor: 상대경로 -> path alias로 전반적인 구조 수정
src directory 제거
